requirements.txt
numpy==1.18.3
Pillow==7.1.2
opencl==2020.1